add funcs

+++ LOCS/backend/static/JS/controllers/adminRouterController.js
@@ -5,6 +5,137 @@ const DataBase = require('../scripts/adminDataBase.js');
 const funcs = require('../scripts/funcs.js');
 const takeObj = funcs.takeObj;
 
+//адрес
+exports.getAddress = async function(request, response) {
+    try {
+        const userId = request.cookies.userId ? await takeObj(request.cookies.userId).then(function(val) { return val.taketoken; }) : undefined;
+        if (userId) {
+            const Role = request.cookies.userRole ? await takeObj(request.cookies.userRole).then(function(val) { return val.taketoken; }) : undefined;
+            if (Role == 0) {
+                await DataBase.getaddress().then(function(val) {
+                    var tags = []
+                    for (i in val) {
+                        tags.push(val[i].getaddress);
+                    }
+                    response.json(tags);
+                }).catch(function(val) {
+                    check = false;
+                    console.log(val);
+                    response.status(500).end(val);
+                });
+            } else {
+                response.status(403).end("have not permissions");
+            }
+        } else {
+            response.status(401).end();
+        }
+    } catch (err) {
+        response.status(500).end(err);
+    }
+
+};
+
+exports.createAddress = async function(request, response) {
+    try {
+        const userId = request.cookies.userId ? await takeObj(request.cookies.userId).then(function(val) { return val.taketoken; }) : undefined;
+        if (userId) {
+            const Role = request.cookies.userRole ? await takeObj(request.cookies.userRole).then(function(val) { return val.taketoken; }) : undefined;
+            if (Role == 0) {
+                let street = request.body.street;
+                let house = request.body.house;
+                let latitude = request.body.latitude;
+                let longitude = request.body.longitude;
+                let id_district = request.body.id_district;
+                let deleted = request.body.deleted;
+                if (street == null || house == null || latitude == null || longitude == null || id_district == null || deleted == null) {
+                    response.status(400).end();
+                    return;
+                }
+                await DataBase.addAddress(street, house, latitude, longitude, id_district, deleted).then(function(val) {
+                    response.json(val);
+                }).catch(function(val) {
+                    check = false;
+                    console.log(val);
+                    response.status(500).end(val);
+                });
+            } else {
+                response.status(403).end("have not permissions");
+            }
+        } else {
+            response.status(401).end();
+        }
+    } catch (err) {
+        response.status(500).end(err);
+    }
+
+};
+
+exports.changeAddress = async function(request, response) {
+    try {
+        const userId = request.cookies.userId ? await takeObj(request.cookies.userId).then(function(val) { return val.taketoken; }) : undefined;
+        if (userId) {
+            const Role = request.cookies.userRole ? await takeObj(request.cookies.userRole).then(function(val) { return val.taketoken; }) : undefined;
+            if (Role == 0) {
+                let id = Number(request.params.id);
+                let street = request.body.street;
+                let house = request.body.house;
+                let latitude = request.body.latitude;
+                let longitude = request.body.longitude;
+                let id_district = request.body.id_district;
+                let deleted = request.body.deleted;
+                if (id == null || street == null || house == null || latitude == null || longitude == null || id_district == null || deleted == null) {
+                    response.status(400).end();
+                    return;
+                }
+                await DataBase.updateAddress(id, street, house, latitude, longitude, id_district, deleted).then(function(val) {
+                    response.json(val);
+                }).catch(function(val) {
+                    check = false;
+                    console.log(val);
+                    response.status(500).end(val);
+                });
+            } else {
+                response.status(403).end("have not permissions");
+            }
+        } else {
+            response.status(401).end();
+        }
+    } catch (err) {
+        response.status(500).end(err);
+    }
+
+};
+
+exports.deleteAddress = async function(request, response) {
+    try {
+        const userId = request.cookies.userId ? await takeObj(request.cookies.userId).then(function(val) { return val.taketoken; }) : undefined;
+        if (userId) {
+            const Role = request.cookies.userRole ? await takeObj(request.cookies.userRole).then(function(val) { return val.taketoken; }) : undefined;
+            if (Role == 0) {
+                let id = Number(request.params.id);
+                if (id == null) {
+                    response.status(400).end();
+                    return;
+                }
+                await DataBase.deleteAddress(id).then(function(val) {
+                    tags = val;
+                    response.status(200).end("deleted");
+                }).catch(function(val) {
+                    check = false;
+                    console.log(val);
+                    response.status(500).end(val);
+                });
+            } else {
+                response.status(403).end("have not permissions");
+            }
+        } else {
+            response.status(401).end();
+        }
+    } catch (err) {
+        response.status(500).end(err);
+    }
+
+};
 //район
 exports.getDistrict = async function(request, response) {
     try {
@@ -44,8 +175,76 @@ exports.createDistrict = async function(request, response) {
                 let title = request.body.title;
                 let id_city = request.body.id_city;
                 let deleted = request.body.deleted;
+                if (title == null || id_city == null || deleted == null) {
+                    response.status(400).end();
+                    return;
+                }
                 await DataBase.addDistrict(title, id_city, deleted).then(function(val) {
                     response.json(val);
+                }).catch(function(val) {
+                    check = false;
+                    console.log(val);
+                    response.status(500).end(val);
+                });
+            } else {
+                response.status(403).end("have not permissions");
+            }
+        } else {
+            response.status(401).end();
+        }
+    } catch (err) {
+        response.status(500).end(err);
+    }
+
+};
+
+exports.changeDistrict = async function(request, response) {
+    try {
+        const userId = request.cookies.userId ? await takeObj(request.cookies.userId).then(function(val) { return val.taketoken; }) : undefined;
+        if (userId) {
+            const Role = request.cookies.userRole ? await takeObj(request.cookies.userRole).then(function(val) { return val.taketoken; }) : undefined;
+            if (Role == 0) {
+                let id = Number(request.params.id);
+                let title = request.body.title;
+                let id_city = request.body.id_city;
+                let deleted = request.body.deleted;
+                if (id == null || title == null || id_city == null || deleted == null) {
+                    response.status(400).end();
+                    return;
+                }
+                await DataBase.updateDistrict(id, title, id_city, deleted).then(function(val) {
+                    response.json(val);
+                }).catch(function(val) {
+                    check = false;
+                    console.log(val);
+                    response.status(500).end(val);
+                });
+            } else {
+                response.status(403).end("have not permissions");
+            }
+        } else {
+            response.status(401).end();
+        }
+    } catch (err) {
+        response.status(500).end(err);
+    }
+
+};
+
+exports.deleteDistrict = async function(request, response) {
+    try {
+        const userId = request.cookies.userId ? await takeObj(request.cookies.userId).then(function(val) { return val.taketoken; }) : undefined;
+        if (userId) {
+            const Role = request.cookies.userRole ? await takeObj(request.cookies.userRole).then(function(val) { return val.taketoken; }) : undefined;
+            if (Role == 0) {
+                let id = Number(request.params.id);
+                if (id == null) {
+                    response.status(400).end();
+                    return;
+                }
+                await DataBase.deleteDistrict(id).then(function(val) {
+                    tags = val;
+                    response.status(200).end("deleted");
                 }).catch(function(val) {
                     check = false;
                     console.log(val);
@@ -101,6 +300,10 @@ exports.createCity = async function(request, response) {
             if (Role == 0) {
                 let title = request.body.title;
                 let deleted = request.body.deleted;
+                if (title == null || deleted == null) {
+                    response.status(400).end();
+                    return;
+                }
                 await DataBase.addCity(title, deleted).then(function(val) {
                     response.json(val);
                 }).catch(function(val) {
@@ -129,6 +332,10 @@ exports.changeCity = async function(request, response) {
                 let id = Number(request.params.id);
                 let title = request.body.title;
                 let deleted = request.body.deleted;
+                if (id == null || title == null || deleted == null) {
+                    response.status(400).end();
+                    return;
+                }
                 await DataBase.updateCity(id, title, deleted).then(function(val) {
                     response.json(val);
                 }).catch(function(val) {
@@ -155,6 +362,10 @@ exports.deleteCity = async function(request, response) {
             const Role = request.cookies.userRole ? await takeObj(request.cookies.userRole).then(function(val) { return val.taketoken; }) : undefined;
             if (Role == 0) {
                 let id = Number(request.params.id);
+                if (id == null) {
+                    response.status(400).end();
+                    return;
+                }
                 await DataBase.deleteCity(id).then(function(val) {
                     response.status(200).end("deleted");
                 }).catch(function(val) {
@@ -185,6 +396,10 @@ exports.changeTag = async function(request, response) {
                 let deleted = request.body.deleted;
                 let accept = request.body.accept;
                 let countevents = request.body.countevents;
+                if (id == null || countevents == null || title == null || accept == null || deleted == null) {
+                    response.status(400).end();
+                    return;
+                }
                 await DataBase.updateTagsAdmin(id, title, deleted, accept, countevents).then(function(val) {
                     response.json(val);
                 }).catch(function(val) {
@@ -215,6 +430,10 @@ exports.addTag = async function(request, response) {
                 let deleted = request.body.deleted;
                 let accept = request.body.accept;
                 let countevents = request.body.countevents;
+                if (countevents == null || title == null || accept == null || deleted == null) {
+                    response.status(400).end();
+                    return;
+                }
                 await DataBase.addTagsAdmin(title, deleted, accept, countevents).then(function(val) {
                     response.json(val);
                 }).catch(function(val) {
@@ -265,14 +484,17 @@ exports.getTag = async function(request, response) {
 };
 
 exports.deleteTag = async function(request, response) {
-
     try {
         const userId = request.cookies.userId ? await takeObj(request.cookies.userId).then(function(val) { return val.taketoken; }) : undefined;
         if (userId) {
             const Role = request.cookies.userRole ? await takeObj(request.cookies.userRole).then(function(val) { return val.taketoken; }) : undefined;
             if (Role == 0) {
-                let idTag = Number(request.params.id);
-                await DataBase.deleteTag(idTag).then(function(val) {
+                let id = Number(request.params.id);
+                if (id == null) {
+                    response.status(400).end();
+                    return;
+                }
+                await DataBase.deleteTag(id).then(function(val) {
                     tags = val;
                     response.status(200).end("deleted");
                 }).catch(function(val) {
@@ -299,6 +521,10 @@ exports.acceptTag = async function(request, response) {
             const Role = request.cookies.userRole ? await takeObj(request.cookies.userRole).then(function(val) { return val.taketoken; }) : undefined;
             if (Role == 0) {
                 let idTag = Number(request.params.id);
+                if (id == null) {
+                    response.status(400).end();
+                    return;
+                }
                 await DataBase.acceptTag(idTag).then(function(val) {
                     tags = val;
                     response.status(200).end("accept");
